Added multiple function for appending/creating user database, logging in and reading from the user database to check credentials.
Regarding the login function, I know we talked about implementing a
userLL generation in there, but I am making a separate function for
that (I feel like it will be simpler) which we can call outside of
login.

Create user appends one user to the database. Not sure how we’re gonna
go about deleting users, but we can figure that out later if need be.

Cipher function is included here because it is necessary wherever
password strings need to be encrypted.