New functions: Create User LL and Append User LL
We have to change the length of the user structure passwords to be
3*MAX_PASSWORD_LENGTH because the cipher lengthens the passwords by a
factor of 3.

I have made a createUserLL function that takes the database, and makes
a LL from it.

I also created an appendUserLL function that adds a new node and also
saves the new node to the database (without overwriting it).